refactor: split HelmCore extension methods for swiftlint
