+++ apps/macos-ui/Helm/Core/HelmCore+Dashboard.swift
@@ -1,4 +1,7 @@
 import Foundation
+import os.log
+
+private let taskSyncLogger = Logger(subsystem: "com.jasoncavinder.Helm", category: "core.tasks")
 
 extension HelmCore {
     var allKnownPackages: [PackageItem] {
@@ -152,6 +155,209 @@ extension HelmCore {
             return .attention
         }
         return .healthy
+    }
+
+    func canUpgradeIndividually(_ package: PackageItem) -> Bool {
+        return package.status == .upgradable
+            && (managerStatuses[package.managerId]?.supportsPackageUpgrade ?? false)
+            && !package.pinned
+            && isManagerEnabled(package.managerId)
+    }
+
+    func canInstallPackage(_ package: PackageItem) -> Bool {
+        package.status == .available
+            && (managerStatuses[package.managerId]?.supportsPackageInstall ?? false)
+            && isManagerEnabled(package.managerId)
+    }
+
+    func canUninstallPackage(_ package: PackageItem) -> Bool {
+        package.status != .available
+            && (managerStatuses[package.managerId]?.supportsPackageUninstall ?? false)
+            && isManagerEnabled(package.managerId)
+    }
+
+    func canPinPackage(_ package: PackageItem) -> Bool {
+        package.status != .available && package.managerId == "homebrew_formula"
+    }
+
+    func supportsRemoteSearch(managerId: String) -> Bool {
+        managerStatuses[managerId]?.supportsRemoteSearch ?? false
+    }
+
+    func isManagerEnabled(_ managerId: String) -> Bool {
+        managerStatuses[managerId]?.enabled ?? true
+    }
+
+    func syncManagerOperations(from coreTasks: [CoreTaskRecord]) {
+        let statusById = Dictionary(uniqueKeysWithValues: coreTasks.map { ($0.id, $0.status.lowercased()) })
+        let inFlightStates = Set(["queued", "running"])
+
+        for managerId in Array(managerActionTaskByManager.keys) {
+            guard let taskId = managerActionTaskByManager[managerId] else { continue }
+            guard let status = statusById[taskId] else {
+                managerOperations.removeValue(forKey: managerId)
+                managerActionTaskByManager.removeValue(forKey: managerId)
+                continue
+            }
+            if !inFlightStates.contains(status) {
+                managerOperations.removeValue(forKey: managerId)
+                managerActionTaskByManager.removeValue(forKey: managerId)
+            }
+        }
+    }
+
+    func syncUpgradeActions(from coreTasks: [CoreTaskRecord]) {
+        let statusById = Dictionary(uniqueKeysWithValues: coreTasks.map { ($0.id, $0.status.lowercased()) })
+        let inFlightStates = Set(["queued", "running"])
+        var shouldRefreshSnapshots = false
+
+        for packageId in Array(upgradeActionTaskByPackage.keys) {
+            guard let taskId = upgradeActionTaskByPackage[packageId] else { continue }
+            guard let status = statusById[taskId] else {
+                upgradeActionTaskByPackage.removeValue(forKey: packageId)
+                upgradeActionPackageIds.remove(packageId)
+                continue
+            }
+            if inFlightStates.contains(status) {
+                continue
+            }
+
+            upgradeActionTaskByPackage.removeValue(forKey: packageId)
+            upgradeActionPackageIds.remove(packageId)
+            if status == "completed" {
+                shouldRefreshSnapshots = true
+            }
+        }
+
+        if shouldRefreshSnapshots {
+            fetchPackages()
+            fetchOutdatedPackages()
+        }
+    }
+
+    func syncInstallActions(from coreTasks: [CoreTaskRecord]) {
+        let statusById = Dictionary(uniqueKeysWithValues: coreTasks.map { ($0.id, $0.status.lowercased()) })
+        let inFlightStates = Set(["queued", "running"])
+        var shouldRefreshSnapshots = false
+
+        for packageId in Array(installActionTaskByPackage.keys) {
+            guard let taskId = installActionTaskByPackage[packageId] else { continue }
+            guard let status = statusById[taskId] else {
+                installActionTaskByPackage.removeValue(forKey: packageId)
+                installActionPackageIds.remove(packageId)
+                continue
+            }
+            if inFlightStates.contains(status) {
+                continue
+            }
+
+            installActionTaskByPackage.removeValue(forKey: packageId)
+            installActionPackageIds.remove(packageId)
+            if status == "completed" {
+                shouldRefreshSnapshots = true
+            }
+        }
+
+        if shouldRefreshSnapshots {
+            fetchPackages()
+            fetchOutdatedPackages()
+            refreshCachedAvailablePackages()
+        }
+    }
+
+    func syncUninstallActions(from coreTasks: [CoreTaskRecord]) {
+        let statusById = Dictionary(uniqueKeysWithValues: coreTasks.map { ($0.id, $0.status.lowercased()) })
+        let inFlightStates = Set(["queued", "running"])
+        var shouldRefreshSnapshots = false
+
+        for packageId in Array(uninstallActionTaskByPackage.keys) {
+            guard let taskId = uninstallActionTaskByPackage[packageId] else { continue }
+            guard let status = statusById[taskId] else {
+                uninstallActionTaskByPackage.removeValue(forKey: packageId)
+                uninstallActionPackageIds.remove(packageId)
+                continue
+            }
+            if inFlightStates.contains(status) {
+                continue
+            }
+
+            uninstallActionTaskByPackage.removeValue(forKey: packageId)
+            uninstallActionPackageIds.remove(packageId)
+            if status == "completed" {
+                shouldRefreshSnapshots = true
+            }
+        }
+
+        if shouldRefreshSnapshots {
+            fetchPackages()
+            fetchOutdatedPackages()
+            refreshCachedAvailablePackages()
+        }
+    }
+
+    func syncPackageDescriptionLookups(from coreTasks: [CoreTaskRecord]) {
+        let statusById = Dictionary(uniqueKeysWithValues: coreTasks.map { ($0.id, $0.status.lowercased()) })
+        let inFlightStates = Set(["queued", "running"])
+
+        for packageId in Array(descriptionLookupTaskIdsByPackage.keys) {
+            guard let taskIds = descriptionLookupTaskIdsByPackage[packageId], !taskIds.isEmpty else {
+                descriptionLookupTaskIdsByPackage.removeValue(forKey: packageId)
+                packageDescriptionLoadingIds.remove(packageId)
+                continue
+            }
+
+            let inFlightTaskIds = taskIds.filter { taskId in
+                guard let status = statusById[taskId] else { return false }
+                return inFlightStates.contains(status)
+            }
+
+            if !inFlightTaskIds.isEmpty {
+                descriptionLookupTaskIdsByPackage[packageId] = inFlightTaskIds
+                continue
+            }
+
+            descriptionLookupTaskIdsByPackage.removeValue(forKey: packageId)
+            packageDescriptionLoadingIds.remove(packageId)
+
+            let summary = allKnownPackages
+                .first(where: { $0.id == packageId })?
+                .summary?
+                .trimmingCharacters(in: .whitespacesAndNewlines)
+
+            if summary?.isEmpty == false {
+                packageDescriptionUnavailableIds.remove(packageId)
+            } else {
+                packageDescriptionUnavailableIds.insert(packageId)
+            }
+        }
+    }
+
+    func updateOnboardingDetectionProgress(from coreTasks: [CoreTaskRecord]) {
+        guard onboardingDetectionInProgress else { return }
+
+        let terminalStatuses = Set(["completed", "failed", "cancelled"])
+        for task in coreTasks
+        where task.id > onboardingDetectionAnchorTaskId
+            && task.taskType.lowercased() == "detection"
+            && terminalStatuses.contains(task.status.lowercased())
+        {
+            onboardingDetectionPendingManagers.remove(task.manager)
+        }
+
+        pruneOnboardingDetectionForDisabledManagers()
+
+        if onboardingDetectionPendingManagers.isEmpty {
+            completeOnboardingDetectionProgress()
+            return
+        }
+
+        if let startedAt = onboardingDetectionStartedAt,
+           Date().timeIntervalSince(startedAt) > 90
+        {
+            let pending = onboardingDetectionPendingManagers.joined(separator: ",")
+            taskSyncLogger.warning("Onboarding detection timed out waiting for managers: \(pending)")
+            completeOnboardingDetectionProgress()
+        }
     }
 
     private func mergeSummary(into package: inout PackageItem, from candidate: String?) {
